feat(user): rename usecases to List and Get
