+++ application/usecases/dealer/GetDealerUsecase.ts
